\name{gen3sdk-package}
\alias{gen3sdk-package}
\alias{gen3sdk}
\docType{package}
\title{
\packageTitle{gen3sdk}
}
\description{
\packageDescription{gen3sdk}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{gen3sdk}
\packageIndices{gen3sdk}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{gen3sdk}

Maintainer: \packageMaintainer{gen3sdk}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
